Update expected files for A4
dorng uses the combined multiple-recursive
generator from L'Ecuyer to produce reproducible
results in parallel environments.
The master seed is used to generate a new random seed
for each value in the foreach loop. That new seed
is used for L'Ecuyer method within each loop. This
only depends on the master seed but it's independent
of the number of cores, only in the order of the loops.